refactor(test-support): consolidate shared test fixtures
