+++ crates/api/Cargo.toml
@@ -16,11 +16,11 @@ uuid = { version = "1.11.0", features = ["v4"] }
 [dev-dependencies]
 async-stream = "0.3.6"
 axum = "0.8.6"
-base64 = "0.22.1"
 bytes = "1.10.1"
 http = "1.3.1"
 selvedge-config = { path = "../config" }
 selvedge-logging = { path = "../logging" }
+selvedge-test-support = { path = "../test-support", features = ["chatgpt-auth", "config", "http"] }
 tempfile = "3.23.0"
 tokio = { version = "1.42.0", features = ["macros", "net", "rt-multi-thread", "sync", "time"] }
