+++ crates/chatgpt-auth/Cargo.toml
@@ -19,6 +19,7 @@ tokio = { version = "1.48.0", features = ["sync"] }
 axum = "0.8.6"
 base64 = "0.22.1"
 selvedge-logging = { path = "../logging" }
+selvedge-test-support = { path = "../test-support", features = ["chatgpt-auth", "config", "http"] }
 tempfile = "3.23.0"
 tokio = { version = "1.48.0", features = ["macros", "net", "rt-multi-thread", "sync", "time"] }
