+++ crates/chatgpt-login/Cargo.toml
@@ -20,6 +20,7 @@ axum = "0.8.6"
 base64 = "0.22.1"
 serde_json = "1.0.145"
 selvedge-logging = { path = "../logging" }
+selvedge-test-support = { path = "../test-support", features = ["config", "http"] }
 tempfile = "3.23.0"
 tokio = { version = "1.48.0", features = ["macros", "net", "rt-multi-thread", "time"] }
