+++ crates/client/Cargo.toml
@@ -21,6 +21,7 @@ zstd = "0.13.3"
 [dev-dependencies]
 axum = "0.8.6"
 rcgen = "0.14.5"
+selvedge-test-support = { path = "../test-support", features = ["config", "http"] }
 tempfile = "3.23.0"
 tokio = { version = "1.48.0", features = ["io-util", "macros", "net", "rt-multi-thread", "sync", "time"] }
 tokio-rustls = "0.26.4"
