+++ crates/core/Cargo.toml
@@ -11,6 +11,7 @@ tokio = { version = "1.42.0", features = ["macros", "rt", "sync"] }
 uuid = { version = "1.11.0", features = ["v4"] }
 
 [dev-dependencies]
+selvedge-test-support = { path = "../test-support", features = ["db-fixtures"] }
 tokio = { version = "1.42.0", features = ["macros", "rt", "sync", "time"] }
 
 [lints.rust]
