+++ crates/local-client/tests/local_client_contract.rs
@@ -1,23 +1,25 @@
-use std::collections::VecDeque;
 use std::fs;
-use std::future;
-use std::pin::Pin;
 use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::{Arc, LazyLock, Mutex};
-use std::task::{Context, Poll};
+use std::sync::{Arc, LazyLock};
 use std::time::Duration;
 
 use futures_util::StreamExt;
 use futures_util::stream;
 use selvedge_local_client::{
     AttachRejectedOrClientError, LocalClientConfig, LocalClientError, LocalClientState,
-    LocalEndpoint, LocalFrameStream, LocalTransport, connect, connect_http,
+    LocalEndpoint, connect, connect_http,
 };
 use selvedge_local_protocol::{
     AttachAccepted, AttachRejectReason, AttachRejected, AttachRequest, CommandOutcome,
-    CommandRejectReason, CommandRequest, CommandResponse, LocalClientCommandId, LocalClientFrame,
-    LocalClientId, LocalClientSubscription, LocalDetailLevel, LocalNotice, LocalNoticeLevel,
-    LocalTaskScope, ReadyRequest, ReadyResponse, ReadyState, current_protocol_version,
+    CommandRejectReason, CommandRequest, CommandResponse, LocalClientCommandId, LocalClientId,
+    LocalClientSubscription, LocalTaskScope, ReadyRequest, ReadyResponse, ReadyState,
+    current_protocol_version,
+};
+use selvedge_test_support::local_transport::{
+    AttachAction, CloseAction, CommandAction, DropNotifyingStream,
+    FakeLocalTransport as FakeTransport, FakeTransportState, PollNotifyingStream, ReadyAction,
+    connect_plan_is_some, connected_client, connected_client_with_timeout, install_connect_plan,
+    next_seq, notice_frame, valid_attach, valid_command, valid_config,
 };
 use tokio::io::{AsyncReadExt, AsyncWriteExt};
 use tokio::net::{TcpListener, TcpStream};
@@ -26,10 +28,6 @@ use tokio::task::JoinHandle;
 use tokio::time::timeout;
 
 static TEST_LOCK: LazyLock<AsyncMutex<()>> = LazyLock::new(|| AsyncMutex::new(()));
-static CONNECT_PLAN: LazyLock<Mutex<Option<Result<FakeTransportStateHandle, LocalClientError>>>> =
-    LazyLock::new(|| Mutex::new(None));
-
-type FakeTransportStateHandle = Arc<Mutex<FakeTransportState>>;
 
 #[tokio::test]
 async fn connect_validates_structured_localhost_endpoint_before_transport_connect() {
@@ -1361,23 +1359,6 @@ fn crate_has_no_systemd_dependency() {
     assert!(!manifest.contains("systemd"));
 }
 
-#[derive(Clone)]
-struct FakeTransport {
-    state: FakeTransportStateHandle,
-}
-
-struct FakeTransportState {
-    connected_configs: Vec<LocalClientConfig>,
-    ready_calls: usize,
-    command_calls: usize,
-    attach_calls: usize,
-    close_calls: usize,
-    close_action: CloseAction,
-    ready_responses: VecDeque<ReadyAction>,
-    command_responses: VecDeque<CommandAction>,
-    attach_responses: VecDeque<AttachAction>,
-}
-
 struct HttpContractResponse {
     status_code: u16,
     content_type: &'static str,
@@ -1391,264 +1372,6 @@ struct CapturedHttpRequest {
     content_type: Option<String>,
     accept: Option<String>,
     body: Vec<u8>,
-}
-
-#[derive(Clone, Copy)]
-enum CloseAction {
-    Complete,
-    Hang,
-}
-
-enum ReadyAction {
-    Response(Result<ReadyResponse, LocalClientError>),
-    Hang,
-}
-
-enum CommandAction {
-    Response(Result<CommandResponse, LocalClientError>),
-    WaitForRelease {
-        release_rx: oneshot::Receiver<()>,
-        response: Result<CommandResponse, LocalClientError>,
-    },
-    Hang,
-}
-
-enum AttachAction {
-    Response(Result<(AttachAccepted, LocalFrameStream), AttachRejectedOrClientError>),
-    Hang,
-}
-
-struct DropNotifyingStream {
-    drops: Arc<AtomicUsize>,
-}
-
-struct PollNotifyingStream {
-    polled: Option<oneshot::Sender<()>>,
-}
-
-impl futures_core::Stream for DropNotifyingStream {
-    type Item = Result<LocalClientFrame, LocalClientError>;
-
-    fn poll_next(self: Pin<&mut Self>, _context: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        Poll::Pending
-    }
-}
-
-impl futures_core::Stream for PollNotifyingStream {
-    type Item = Result<LocalClientFrame, LocalClientError>;
-
-    fn poll_next(mut self: Pin<&mut Self>, _context: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        if let Some(polled) = self.polled.take() {
-            let _ = polled.send(());
-        }
-        Poll::Pending
-    }
-}
-
-impl Drop for DropNotifyingStream {
-    fn drop(&mut self) {
-        self.drops.fetch_add(1, Ordering::SeqCst);
-    }
-}
-
-impl FakeTransportState {
-    fn new_handle() -> FakeTransportStateHandle {
-        Arc::new(Mutex::new(Self {
-            connected_configs: Vec::new(),
-            ready_calls: 0,
-            command_calls: 0,
-            attach_calls: 0,
-            close_calls: 0,
-            close_action: CloseAction::Complete,
-            ready_responses: VecDeque::new(),
-            command_responses: VecDeque::new(),
-            attach_responses: VecDeque::new(),
-        }))
-    }
-}
-
-impl LocalTransport for FakeTransport {
-    async fn connect(config: LocalClientConfig) -> Result<Self, LocalClientError> {
-        let state = CONNECT_PLAN
-            .lock()
-            .expect("connect plan lock")
-            .take()
-            .expect("connect plan installed")?;
-        state
-            .lock()
-            .expect("fake state")
-            .connected_configs
-            .push(config);
-        Ok(Self { state })
-    }
-
-    async fn ready(&self, _request: ReadyRequest) -> Result<ReadyResponse, LocalClientError> {
-        let action = {
-            let mut state = self.state.lock().expect("fake state");
-            state.ready_calls += 1;
-            state
-                .ready_responses
-                .pop_front()
-                .unwrap_or(ReadyAction::Response(Ok(ReadyResponse {
-                    protocol_version: current_protocol_version(),
-                    state: ReadyState::Ready,
-                })))
-        };
-
-        match action {
-            ReadyAction::Response(response) => response,
-            ReadyAction::Hang => future::pending().await,
-        }
-    }
-
-    async fn submit_command(
-        &self,
-        _request: CommandRequest,
-    ) -> Result<CommandResponse, LocalClientError> {
-        let action = {
-            let mut state = self.state.lock().expect("fake state");
-            state.command_calls += 1;
-            state.command_responses.pop_front().unwrap_or_else(|| {
-                CommandAction::Response(Ok(CommandResponse {
-                    protocol_version: current_protocol_version(),
-                    client_command_id: LocalClientCommandId::new("command-1").expect("command id"),
-                    outcome: CommandOutcome::Accepted,
-                }))
-            })
-        };
-
-        match action {
-            CommandAction::Response(response) => response,
-            CommandAction::WaitForRelease {
-                release_rx,
-                response,
-            } => {
-                let _ = release_rx.await;
-                response
-            }
-            CommandAction::Hang => future::pending().await,
-        }
-    }
-
-    async fn attach(
-        &self,
-        _request: AttachRequest,
-    ) -> Result<(AttachAccepted, LocalFrameStream), AttachRejectedOrClientError> {
-        let action = {
-            let mut state = self.state.lock().expect("fake state");
-            state.attach_calls += 1;
-            state.attach_responses.pop_front().unwrap_or_else(|| {
-                AttachAction::Response(Ok((
-                    AttachAccepted {
-                        protocol_version: current_protocol_version(),
-                        client_id: LocalClientId::new("client-1").expect("client id"),
-                        client_command_id: LocalClientCommandId::new("attach-1")
-                            .expect("command id"),
-                    },
-                    Box::pin(stream::empty()),
-                )))
-            })
-        };
-
-        match action {
-            AttachAction::Response(response) => response,
-            AttachAction::Hang => future::pending().await,
-        }
-    }
-
-    async fn close(&self) {
-        let action = {
-            let mut state = self.state.lock().expect("fake state");
-            state.close_calls += 1;
-            state.close_action
-        };
-
-        match action {
-            CloseAction::Complete => {}
-            CloseAction::Hang => future::pending().await,
-        }
-    }
-}
-
-async fn connected_client(
-    state: FakeTransportStateHandle,
-) -> selvedge_local_client::LocalClient<FakeTransport> {
-    connected_client_with_timeout(state, Duration::from_secs(1)).await
-}
-
-async fn connected_client_with_timeout(
-    state: FakeTransportStateHandle,
-    request_timeout: Duration,
-) -> selvedge_local_client::LocalClient<FakeTransport> {
-    install_connect_plan(Ok(state));
-    connect::<FakeTransport>(LocalClientConfig {
-        endpoint: LocalEndpoint::TcpIpv4 { port: 17691 },
-        request_timeout,
-    })
-    .await
-    .expect("connect client")
-}
-
-fn install_connect_plan(plan: Result<FakeTransportStateHandle, LocalClientError>) {
-    *CONNECT_PLAN.lock().expect("connect plan lock") = Some(plan);
-}
-
-fn connect_plan_is_some() -> bool {
-    CONNECT_PLAN.lock().expect("connect plan lock").is_some()
-}
-
-fn valid_config() -> LocalClientConfig {
-    LocalClientConfig {
-        endpoint: LocalEndpoint::TcpIpv4 { port: 17691 },
-        request_timeout: Duration::from_secs(1),
-    }
-}
-
-fn valid_command(command_id: &str) -> CommandRequest {
-    CommandRequest {
-        protocol_version: current_protocol_version(),
-        client_id: LocalClientId::new("client-1").expect("client id"),
-        client_command_id: LocalClientCommandId::new(command_id).expect("command id"),
-        command_name: "send-user-input".to_owned(),
-        payload: serde_json::json!({"message": "hello"}),
-    }
-}
-
-fn valid_attach(command_id: &str) -> AttachRequest {
-    AttachRequest {
-        protocol_version: current_protocol_version(),
-        client_id: LocalClientId::new("client-1").expect("client id"),
-        client_command_id: LocalClientCommandId::new(command_id).expect("command id"),
-        subscription: LocalClientSubscription {
-            task_scope: LocalTaskScope::AllTasks,
-            detail_level: LocalDetailLevel::Summary,
-            include_model_call_status: false,
-            include_tool_execution_status: false,
-            include_debug_notices: false,
-        },
-    }
-}
-
-fn notice_frame(seq: u64) -> LocalClientFrame {
-    LocalClientFrame::Notice(selvedge_local_protocol::LocalClientNoticeFrame {
-        delivery_seq: seq,
-        client_command_id: LocalClientCommandId::new(format!("notice-{seq}")).expect("command id"),
-        notice: LocalNotice {
-            level: LocalNoticeLevel::Info,
-            message_text: format!("notice {seq}"),
-        },
-    })
-}
-
-async fn next_seq(stream: &mut LocalFrameStream) -> Result<u64, LocalClientError> {
-    match stream.next().await {
-        Some(Ok(LocalClientFrame::Notice(frame))) => Ok(frame.delivery_seq),
-        Some(Ok(_)) => Err(LocalClientError::TransportFailed(
-            "unexpected frame kind".to_owned(),
-        )),
-        Some(Err(error)) => Err(error),
-        None => Err(LocalClientError::StreamClosed),
-    }
 }
 
 impl HttpContractResponse {
