+++ crates/task-runtime-factory/Cargo.toml
@@ -12,6 +12,7 @@ tokio = { version = "1.42.0", features = ["rt", "sync"] }
 uuid = { version = "1.11.0", features = ["v4"] }
 
 [dev-dependencies]
+selvedge-test-support = { path = "../test-support", features = ["db-fixtures"] }
 tokio = { version = "1.42.0", features = ["macros", "rt", "sync", "time"] }
 
 [lints.rust]
