+++ crates/tui/tests/tui_contract.rs
@@ -1,28 +1,22 @@
-use std::collections::VecDeque;
-use std::future;
-use std::sync::{Arc, LazyLock, Mutex};
+use std::sync::LazyLock;
 use std::time::Duration;
 
-use futures_util::stream;
-use selvedge_local_client::{
-    AttachRejectedOrClientError, LocalClientConfig, LocalClientError, LocalEndpoint,
-    LocalFrameStream, LocalTransport,
-};
+use selvedge_local_client::{LocalClientConfig, LocalClientError, LocalEndpoint};
 use selvedge_local_protocol::{
-    AttachAccepted, AttachRejectReason, AttachRejected, AttachRequest, CommandOutcome,
-    CommandRejectReason, CommandRequest, CommandResponse, LocalClientCommandId, LocalClientFrame,
-    LocalClientId, LocalClientSnapshot, LocalClientSnapshotFrame, LocalClientSubscription,
-    LocalDetailLevel, LocalTaskScope, ReadyRequest, ReadyResponse, ReadyState,
+    AttachRejectReason, AttachRejected, CommandOutcome, CommandRejectReason, CommandRequest,
+    CommandResponse, LocalClientCommandId, LocalClientFrame, LocalClientSnapshotFrame,
+    LocalClientSubscription, LocalDetailLevel, LocalTaskScope, ReadyResponse, ReadyState,
     current_protocol_version,
+};
+use selvedge_test_support::local_transport::{
+    AttachAction, CommandAction, FakeLocalTransport as FakeTransport, FakeTransportState,
+    ReadyAction, empty_local_snapshot as empty_snapshot, install_connect_plan,
+    noop_command as valid_command, ready_state,
 };
 use selvedge_tui::{TuiExitStatus, TuiInputAction, TuiStartArgs, run_tui};
 use tokio::sync::Mutex as AsyncMutex;
 
 static TEST_LOCK: LazyLock<AsyncMutex<()>> = LazyLock::new(|| AsyncMutex::new(()));
-static CONNECT_PLAN: LazyLock<Mutex<Option<Result<FakeTransportStateHandle, LocalClientError>>>> =
-    LazyLock::new(|| Mutex::new(None));
-
-type FakeTransportStateHandle = Arc<Mutex<FakeTransportState>>;
 
 // @verifies selvedge.client.tui.run
 #[tokio::test]
@@ -44,10 +38,10 @@ async fn not_ready_returns_server_not_ready() {
         .lock()
         .expect("fake state")
         .ready_responses
-        .push_back(ReadyResponse {
+        .push_back(ReadyAction::Response(Ok(ReadyResponse {
             protocol_version: current_protocol_version(),
             state: ReadyState::NotReady,
-        });
+        })));
     install_connect_plan(Ok(state.clone()));
 
     let status = run_tui::<FakeTransport, _>(valid_args(None), NoopMapper).await;
@@ -101,11 +95,13 @@ async fn waits_for_snapshot_then_submits_initial_command_and_reports_rejection()
                     snapshot: empty_snapshot(),
                 }),
             )]));
-        state_guard.command_responses.push_back(CommandResponse {
-            protocol_version: current_protocol_version(),
-            client_command_id: LocalClientCommandId::new("command-1").expect("command id"),
-            outcome: CommandOutcome::Rejected(CommandRejectReason::UnsupportedCommand),
-        });
+        state_guard
+            .command_responses
+            .push_back(CommandAction::Response(Ok(CommandResponse {
+                protocol_version: current_protocol_version(),
+                client_command_id: LocalClientCommandId::new("command-1").expect("command id"),
+                outcome: CommandOutcome::Rejected(CommandRejectReason::UnsupportedCommand),
+            })));
     }
     install_connect_plan(Ok(state.clone()));
 
@@ -142,11 +138,13 @@ async fn accepted_initial_command_exits_successfully() {
                     snapshot: empty_snapshot(),
                 }),
             )]));
-        state_guard.command_responses.push_back(CommandResponse {
-            protocol_version: current_protocol_version(),
-            client_command_id: LocalClientCommandId::new("command-1").expect("command id"),
-            outcome: CommandOutcome::Accepted,
-        });
+        state_guard
+            .command_responses
+            .push_back(CommandAction::Response(Ok(CommandResponse {
+                protocol_version: current_protocol_version(),
+                client_command_id: LocalClientCommandId::new("command-1").expect("command id"),
+                outcome: CommandOutcome::Accepted,
+            })));
     }
     install_connect_plan(Ok(state.clone()));
 
@@ -212,138 +210,6 @@ impl selvedge_tui::TuiCommandMapper for NoopMapper {
     }
 }
 
-#[derive(Clone)]
-struct FakeTransport {
-    state: FakeTransportStateHandle,
-}
-
-#[derive(Default)]
-struct FakeTransportState {
-    ready_responses: VecDeque<ReadyResponse>,
-    attach_responses: VecDeque<AttachAction>,
-    command_responses: VecDeque<CommandResponse>,
-    ready_calls: usize,
-    attach_calls: usize,
-    command_calls: usize,
-    close_calls: usize,
-}
-
-enum AttachAction {
-    Accepted(Vec<Result<LocalClientFrame, LocalClientError>>),
-    Rejected(AttachRejected),
-    Pending,
-}
-
-impl FakeTransportState {
-    fn new_handle() -> FakeTransportStateHandle {
-        Arc::new(Mutex::new(Self::default()))
-    }
-}
-
-impl LocalTransport for FakeTransport {
-    async fn connect(config: LocalClientConfig) -> Result<Self, LocalClientError>
-    where
-        Self: Sized,
-    {
-        let _ = config;
-        match CONNECT_PLAN.lock().expect("connect plan").take() {
-            Some(Ok(state)) => Ok(Self { state }),
-            Some(Err(error)) => Err(error),
-            None => Err(LocalClientError::ConnectFailed(
-                "missing connect plan".to_owned(),
-            )),
-        }
-    }
-
-    async fn ready(&self, request: ReadyRequest) -> Result<ReadyResponse, LocalClientError> {
-        let _ = request;
-        let mut state = self.state.lock().expect("fake state");
-        state.ready_calls += 1;
-        Ok(state.ready_responses.pop_front().unwrap_or(ReadyResponse {
-            protocol_version: current_protocol_version(),
-            state: ReadyState::Ready,
-        }))
-    }
-
-    async fn submit_command(
-        &self,
-        request: CommandRequest,
-    ) -> Result<CommandResponse, LocalClientError> {
-        let mut state = self.state.lock().expect("fake state");
-        state.command_calls += 1;
-        Ok(state
-            .command_responses
-            .pop_front()
-            .unwrap_or(CommandResponse {
-                protocol_version: current_protocol_version(),
-                client_command_id: request.client_command_id,
-                outcome: CommandOutcome::Accepted,
-            }))
-    }
-
-    async fn attach(
-        &self,
-        request: AttachRequest,
-    ) -> Result<(AttachAccepted, LocalFrameStream), AttachRejectedOrClientError> {
-        let mut state = self.state.lock().expect("fake state");
-        state.attach_calls += 1;
-        match state.attach_responses.pop_front() {
-            Some(AttachAction::Accepted(frames)) => Ok((
-                AttachAccepted {
-                    protocol_version: current_protocol_version(),
-                    client_id: request.client_id,
-                    client_command_id: request.client_command_id,
-                },
-                Box::pin(stream::iter(frames)),
-            )),
-            Some(AttachAction::Rejected(rejected)) => {
-                Err(AttachRejectedOrClientError::Rejected(rejected))
-            }
-            Some(AttachAction::Pending) => Ok((
-                AttachAccepted {
-                    protocol_version: current_protocol_version(),
-                    client_id: request.client_id,
-                    client_command_id: request.client_command_id,
-                },
-                Box::pin(stream::pending()),
-            )),
-            None => Ok((
-                AttachAccepted {
-                    protocol_version: current_protocol_version(),
-                    client_id: request.client_id,
-                    client_command_id: request.client_command_id,
-                },
-                Box::pin(stream::iter(Vec::new())),
-            )),
-        }
-    }
-
-    fn close(&self) -> impl Future<Output = ()> + Send {
-        let state = self.state.clone();
-        {
-            state.lock().expect("fake state").close_calls += 1;
-        };
-        future::ready(())
-    }
-}
-
-fn install_connect_plan(plan: Result<FakeTransportStateHandle, LocalClientError>) {
-    *CONNECT_PLAN.lock().expect("connect plan") = Some(plan);
-}
-
-fn ready_state() -> FakeTransportStateHandle {
-    let state = FakeTransportState::new_handle();
-    state
-        .lock()
-        .expect("fake state")
-        .ready_responses
-        .push_back(ReadyResponse {
-            protocol_version: current_protocol_version(),
-            state: ReadyState::Ready,
-        });
-    state
-}
-
 fn valid_args(initial_command: Option<CommandRequest>) -> TuiStartArgs {
     TuiStartArgs {
         client_config: LocalClientConfig {
@@ -360,25 +226,5 @@ fn valid_args(initial_command: Option<CommandRequest>) -> TuiStartArgs {
             include_debug_notices: false,
         },
         initial_command,
-    }
-}
-
-fn valid_command(command_id: &str) -> CommandRequest {
-    CommandRequest {
-        protocol_version: current_protocol_version(),
-        client_id: LocalClientId::new("client-1").expect("client id"),
-        client_command_id: LocalClientCommandId::new(command_id).expect("command id"),
-        command_name: "noop".to_owned(),
-        payload: serde_json::json!({}),
-    }
-}
-
-fn empty_snapshot() -> LocalClientSnapshot {
-    LocalClientSnapshot {
-        generated_at: 1,
-        tasks: Vec::new(),
-        task_parent_edges: Vec::new(),
-        history_nodes: Vec::new(),
-        task_versions: Vec::new(),
     }
 }
